chore: 진행 상황 저장 및 fix: TodoInput 컴포넌트 분류
