+++ components/DailyTodos.jsx
@@ -1,23 +1,10 @@
-import { CategoryContext } from '@/contexts/CategoryContext';
-import useTodoStore from '@/contexts/TodoStore';
-import { Input, List } from '@ui-kitten/components';
-import { useContext, useState } from 'react';
+import { List } from '@ui-kitten/components';
 import DailyTodo from './DailyTodo';
+import TodoInput from './TodoInput';
 
 const DailyTodos = ({ todos }) => {
-  const addTodo = useTodoStore(state => state.addTodo);
-
-  const [input, setInput] = useState('');
-  const { selectedCategory } = useContext(CategoryContext);
-
   const renderTodo = ({ item, index }) => {
-    return <DailyTodo item={item} />;
-  };
-  const handleSubmit = async () => {
-    const startDate = new Date().toISOString().split('T')[0];
-    const endDate = new Date().toISOString().split('T')[0];
-    addTodo(startDate, endDate, input, selectedCategory);
-    setInput('');
+    return <DailyTodo item={item} key={index} />;
   };
 
   return (
@@ -25,14 +12,7 @@ const DailyTodos = ({ todos }) => {
       data={todos}
       renderItem={renderTodo}
       contentContainerStyle={{ paddingBottom: 100 }}
-      ListFooterComponent={
-        <Input
-          placeholder="Place your Text"
-          value={input}
-          onChangeText={nextInput => setInput(nextInput)}
-          onSubmitEditing={handleSubmit}
-        />
-      }
+      ListFooterComponent={<TodoInput />}
       ListFooterComponentStyle={{ paddingTop: 0, paddingBottom: 125 }}
     />
   );
