+++ package.json
@@ -33,6 +33,7 @@
     "react-dom": "18.2.0",
     "react-native": "0.74.2",
     "react-native-elements": "^3.4.3",
+    "react-native-keyboard-aware-scroll-view": "^0.9.5",
     "react-native-safe-area-context": "4.10.1",
     "react-native-screens": "3.31.1",
     "react-native-svg": "^9.13.6",
